requirements.txt
# 核心依赖
aiohttp>=3.9.0
feedparser>=6.0.0
python-dotenv>=1.0.0

# 可选：本地测试用
# pytest>=7.0.0
# pytest-asyncio>=0.23.0
